order_services/pyproject.toml
[tool.poetry]
name = "order-services-"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "order_services"}]

[tool.poetry.dependencies]
python = "^3.12"
sqlmodel = "^0.0.22"
fastapi = "^0.115.2"
aiokafka = "^0.11.0"
uvicorn = {extras = ["binary"], version = "^0.31.1"}
psycopg2 = "^2.9.10"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic = {extras = ["email"], version = "^2.10.1"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
